# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Payment gateway aggregator — VA, QRIS, and disbursement via Midtrans, Xendit, DOKU, iPaymu.\n\n## Authentication\nMerchant endpoints (/v1/*): `X-API-Key` header with your API key (`wpay_live_*` or `wpay_test_*`).\nAdmin endpoints (/admin/*): `Authorization: Bearer \u003caccess_token\u003e` from POST /admin/login.\n\n## Idempotency\nPayment and disbursement creation support `Idempotency-Key` header. Use unique key per request to safely retry.\n\n## Webhook\nWanpey signs outbound webhooks with `X-Wanpey-Signature` (HMAC-SHA256). Verify using your webhook secret.",
        "title": "Wanpey Payment Gateway API",
        "contact": {
            "name": "Wanpey Support",
            "email": "support@wanpey.id"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/merchants/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the authenticated merchant's profile, status, and configuration.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "Get merchant profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wanpey_core_internal_usecase.MerchantOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates merchant name, webhook URL, or other profile fields. FeeConfig cannot be changed here — contact admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "Update merchant profile",
                "parameters": [
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_internal_usecase.UpdateMerchantInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wanpey_core_internal_usecase.MerchantOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/merchants/me/api-key/regenerate": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Generates a new API key and immediately invalidates the old one. The raw key is shown only once — store it securely.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "Regenerate API key",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "api_key": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/merchants/me/bank-accounts": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns all registered bank accounts for the authenticated merchant (max 3).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant / Bank Accounts"
                ],
                "summary": "List bank accounts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/wanpey_core_internal_usecase.BankAccountOutput"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Registers a new bank account for disbursement. Max 3 accounts per merchant. New accounts must be verified by admin before they can receive disbursements.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant / Bank Accounts"
                ],
                "summary": "Add bank account",
                "parameters": [
                    {
                        "description": "Bank account details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_internal_usecase.AddBankAccountInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/wanpey_core_internal_usecase.BankAccountOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Maximum 3 bank accounts reached",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/merchants/me/bank-accounts/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Removes a registered bank account. Cannot remove the primary account if other accounts exist.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant / Bank Accounts"
                ],
                "summary": "Remove bank account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bank account ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Bank account removed"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/merchants/me/bank-accounts/{id}/primary": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Marks a bank account as the primary account for disbursements.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant / Bank Accounts"
                ],
                "summary": "Set primary bank account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bank account ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/merchants/me/webhook-events": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns outbound webhook delivery history for this merchant — useful for debugging failed deliveries.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "List webhook delivery events",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/wanpey_core_pkg_response.ListResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/wanpey_core_internal_usecase.WebhookEventOutput"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/wanpey_core_pkg_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "wanpey_core_internal_domain_entity.BankCode": {
            "type": "string",
            "enum": [
                "BCA",
                "BNI",
                "BRI",
                "BSI",
                "MANDIRI",
                "PERMATA",
                "CIMB"
            ],
            "x-enum-varnames": [
                "BankBCA",
                "BankBNI",
                "BankBRI",
                "BankBSI",
                "BankMandiri",
                "BankPermata",
                "BankCIMB"
            ]
        },
...
        "wanpey_core_internal_domain_entity.FeeConfig": {
            "type": "object",
            "properties": {
                "disbursement": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.MethodFee"
                },
                "qris": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.MethodFee"
                },
                "va": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.MethodFee"
                }
            }
        },
...
        "wanpey_core_internal_domain_entity.MerchantStatus": {
            "type": "string",
            "enum": [
                "pending",
                "active",
                "suspended",
                "inactive"
            ],
            "x-enum-comments": {
                "MerchantStatusInactive": "closed account",
                "MerchantStatusPending": "awaiting KYC/approval",
                "MerchantStatusSuspended": "temporarily blocked"
            },
            "x-enum-descriptions": [
                "awaiting KYC/approval",
                "",
                "temporarily blocked",
                "closed account"
            ],
            "x-enum-varnames": [
                "MerchantStatusPending",
                "MerchantStatusActive",
                "MerchantStatusSuspended",
                "MerchantStatusInactive"
            ]
        },
...
        "wanpey_core_internal_usecase.AddBankAccountInput": {
            "type": "object",
            "required": [
                "account_name",
                "account_number",
                "bank_code"
            ],
            "properties": {
                "account_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "account_number": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 5
                },
                "bank_code": {
                    "enum": [
                        "BCA",
                        "BNI",
                        "BRI",
                        "BSI",
                        "MANDIRI",
                        "PERMATA",
                        "CIMB"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/wanpey_core_internal_domain_entity.BankCode"
                        }
                    ]
                },
                "set_as_primary": {
                    "type": "boolean"
                }
            }
        },
...
        "wanpey_core_internal_usecase.BankAccountOutput": {
            "type": "object",
            "properties": {
                "account_name": {
                    "type": "string"
                },
                "account_number": {
                    "type": "string"
                },
                "bank_code": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.BankCode"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_primary": {
                    "type": "boolean"
                },
                "is_verified": {
                    "type": "boolean"
                }
            }
        },
...
        "wanpey_core_internal_usecase.MerchantOutput": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fee_config": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.FeeConfig"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/wanpey_core_internal_domain_entity.MerchantStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "webhook_url": {
                    "type": "string"
                }
            }
        },
...
        "wanpey_core_internal_usecase.UpdateMerchantInput": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "name": {
                    "type": "string",
                    "maxLength": 100
                },
                "phone": {
                    "type": "string",
                    "maxLength": 20
                },
                "webhook_url": {
                    "type": "string",
                    "maxLength": 500
                }
            }
        },
        "wanpey_core_internal_usecase.WebhookEventOutput": {
            "type": "object",
            "properties": {
                "attempt_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "delivered_at": {
                    "type": "string"
                },
                "event_type": {
                    "type": "string"
                },
                "failed_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "next_retry_at": {
                    "type": "string"
                },
                "target_url": {
                    "type": "string"
                }
            }
        },
        "wanpey_core_pkg_response.ErrorBody": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/wanpey_core_pkg_response.FieldDetail"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "wanpey_core_pkg_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.ErrorBody"
                },
                "meta": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "wanpey_core_pkg_response.ListResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "wanpey_core_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Pagination"
                },
                "timestamp": {
                    "type": "string"
                },
                "trace_id": {
                    "type": "string"
                }
            }
        },
        "wanpey_core_pkg_response.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "wanpey_core_pkg_response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/wanpey_core_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin JWT — Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ApiKeyAuth": {
            "description": "Merchant API key",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}